add properties files to load data for roles
this is very specific right now, the file has to be data.sql
as opposed to schema.sql. And it has to be at the top-level
of the resources directory. Must be some other way to be
more configurable, but for now this is the default config that work.
Also, using create/destroy to ensure the loading is working